requirements.txt
python-telegram-bot
requests
pandas
numpy
scikit-learn
xgboost
matplotlib
python-dotenv
